Crate google_testing1

This documentation was generated from testing crate version 1.0.9+20190627, where 20190627 is the exact revision of the testing:v1 schema built by the mako code generator v1.0.9.

Everything else about the testing v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

application detail service

get apk details

projects

test matrices cancel, test matrices create and test matrices get

test environment catalog

get

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().test_matrices_create(...).doit()
let r = hub.projects().test_matrices_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-testing1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_testing1 as testing1;
use testing1::TestMatrix;
use testing1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use testing1::Testing;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Testing::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = TestMatrix::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().test_matrices_create(req, "projectId")
             .request_id("sed")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Account
Identifies an account and how to log into it.

AndroidDevice
A single Android device.

AndroidDeviceCatalog
The currently supported Android devices.

AndroidDeviceList
A list of Android device configurations in which the test is to be executed.

AndroidInstrumentationTest
A test of an Android application that can control an Android component independently of its normal lifecycle. Android instrumentation tests run an application APK and test APK inside the same process on a virtual or physical AndroidDevice. They also specify a test runner class, such as com.google.GoogleTestRunner, which can vary on the specific instrumentation framework chosen.

AndroidMatrix
A set of Android device configuration permutations is defined by the the cross-product of the given axes. Internally, the given AndroidMatrix will be expanded into a set of AndroidDevices.

AndroidModel
A description of an Android device tests may be run on.

AndroidRoboTest
A test of an android application that explores the application on a virtual or physical Android Device, finding culprits and crashes as it goes.

AndroidRuntimeConfiguration
Android configuration that can be selected at the time a test is run.

AndroidTestLoop
A test of an Android Application with a Test Loop. The intent <intent-name> will be implicitly added, since Games is the only user of this api, for the time being.

AndroidVersion
A version of the Android OS.

Apk
An Android package file to install.

ApkDetail
Android application details based on application manifest and apk archive contents.

ApkManifest
An Android app manifest. See http://developer.android.com/guide/topics/manifest/manifest-intro.html

AppBundle
An Android App Bundle file format, containing a BundleConfig.pb file, a base module directory, zero or more dynamic feature module directories.

ApplicationDetailServiceGetApkDetailCall
Gets the details of an Android application APK.

ApplicationDetailServiceMethods
A builder providing access to all methods supported on applicationDetailService resources. It is not used directly, but through the Testing hub.

CancelTestMatrixResponse
Response containing the current state of the specified test matrix.

Chunk

ClientInfo
Information about the client which invoked the test.

ClientInfoDetail
Key-value pair of detailed information about the client which invoked the test. Examples: {'Version', '1.0'}, {'Release Track', 'BETA'}.

ContentRange
Implements the Content-Range header, for serialization only

Date
Represents a whole or partial calendar date, e.g. a birthday. The time of day and time zone are either specified elsewhere or are not significant. The date is relative to the Proleptic Gregorian Calendar. This can represent:

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DeviceFile
A single device file description.

Distribution
Data about the relative number of devices running a given configuration of the Android platform.

DummyNetworkStream

Environment
The environment in which the test is run.

EnvironmentMatrix
The matrix of environments in which the test is to be executed.

EnvironmentVariable
A key-value pair passed as an environment variable to the test.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

FileReference
A reference to a file, used for user inputs.

GetApkDetailsResponse
Response containing the details of the specified Android application APK.

GoogleAuto
Enables automatic Google account login. If set, the service automatically generates a Google test account and adds it to the device, before executing the test. Note that test accounts might be reused. Many applications show their full set of functionalities when an account is present on the device. Logging into the device with these generated accounts allows testing more functionalities.

GoogleCloudStorage
A storage location within Google cloud storage (GCS).

IntentFilter
The section of an tag. https://developer.android.com/guide/topics/manifest/intent-filter-element.html

IosDevice
A single iOS device.

IosDeviceCatalog
The currently supported iOS devices.

IosDeviceList
A list of iOS device configurations in which the test is to be executed.

IosModel
A description of an iOS device tests may be run on.

IosRuntimeConfiguration
iOS configuration that can be selected at the time a test is run.

IosTestSetup
A description of how to set up an iOS device prior to running the test.

IosVersion
An iOS version.

IosXcTest
A test of an iOS application that uses the XCTest framework. Xcode supports the option to "build for testing", which generates an .xctestrun file that contains a test specification (arguments, test methods, etc). This test type accepts a zip file containing the .xctestrun file and the corresponding contents of the Build/Products directory that contains all the binaries needed to run the tests.

JsonServerError
A utility type which can decode a server response that indicates error

LauncherActivityIntent
Specifies an intent that starts the main launcher activity.

Locale
A location/region designation for language.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

NetworkConfiguration
There is no detailed description.

NetworkConfigurationCatalog
There is no detailed description.

ObbFile
An opaque binary blob file to install on the device before the test starts.

Orientation
Screen orientation of the device.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Testing hub.

ProjectTestMatriceCancelCall
Cancels unfinished test executions in a test matrix. This call returns immediately and cancellation proceeds asychronously. If the matrix is already final, this operation will have no effect.

ProjectTestMatriceCreateCall
Creates and runs a matrix of tests according to the given specifications. Unsupported environments will be returned in the state UNSUPPORTED. Matrices are limited to at most 200 supported executions.

ProjectTestMatriceGetCall
Checks the status of a test matrix.

ProvidedSoftwareCatalog
The currently provided software environment on the devices under test.

RangeResponseHeader

RegularFile
A file or directory to install on the device before the test starts.

ResultStorage
Locations where the results of running the test are stored.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

RoboDirective
Directs Robo to interact with a specific UI element if it is encountered during the crawl. Currently, Robo can perform text entry or element click.

RoboStartingIntent
Message for specifying the start activities to crawl.

ServerError

ServerMessage

StartActivityIntent
A starting intent specified by an action, uri, and categories.

TestDetails
Additional details about the progress of the running test.

TestEnvironmentCatalog
A description of a test environment.

TestEnvironmentCatalogGetCall
Gets the catalog of supported test environments.

TestEnvironmentCatalogMethods
A builder providing access to all methods supported on testEnvironmentCatalog resources. It is not used directly, but through the Testing hub.

TestExecution
A single test executed in a single environment.

TestMatrix
TestMatrix captures all details about a test. It contains the environment configuration, test specification, test executions and overall state and outcome.

TestSetup
A description of how to set up the Android device prior to running the test.

TestSpecification
A description of how to run the test.

Testing
Central instance to access all Testing related resource activities

ToolResultsExecution
Represents a tool results execution resource.

ToolResultsHistory
Represents a tool results history resource.

ToolResultsStep
Represents a tool results step resource.

TrafficRule
Network emulation parameters.

XUploadContentType
The X-Upload-Content-Type header.

XcodeVersion
An Xcode version that an iOS version is compatible with.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.